update to snapshot of upcoming aepp-sdk-java
