+++ provider-aeternity/pom.xml
@@ -13,7 +13,7 @@
   <artifactId>provider-aeternity</artifactId>
 
   <properties>
-    <aepp-sdk-java.version>1.2.0</aepp-sdk-java.version>
+    <aepp-sdk-java.version>1.2.1-SNAPSHOT</aepp-sdk-java.version>
   </properties>
 
   <repositories>
@@ -22,8 +22,8 @@
       <url>https://jcenter.bintray.com/</url>
     </repository>
     <repository>
-      <id>tuweni</id>
-      <url>https://dl.bintray.com/tuweni/tuweni/</url>
+      <id>oss-snapshot-local</id>
+      <url>https://oss.jfrog.org/artifactory/oss-snapshot-local</url>
     </repository>
   </repositories>
