+++ provider-aeternity/src/main/java/network/arkane/provider/aeternity/AeternityServiceAutoConfiguration.java
@@ -1,0 +1,39 @@
+package network.arkane.provider.aeternity;
+
+import com.kryptokrauts.aeternity.sdk.constants.Network;
+import com.kryptokrauts.aeternity.sdk.constants.VirtualMachine;
+import com.kryptokrauts.aeternity.sdk.service.aeternity.AeternityServiceConfiguration;
+import com.kryptokrauts.aeternity.sdk.service.aeternity.AeternityServiceFactory;
+import com.kryptokrauts.aeternity.sdk.service.aeternity.impl.AeternityService;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class AeternityServiceAutoConfiguration {
+
+    @Value("${network.arkane.aeternity.node.api.baseUrl:https://sdk-testnet.aepps.com/v2}")
+    String baseUrl;
+    @Value("${network.arkane.aeternity.compiler.api.baseUrl:https://compiler.aepps.com}")
+    String compilerBaseUrl;
+    @Value("${network.arkane.aeternity.network:TESTNET}")
+    Network network;
+    @Value("${network.arkane.aeternity.vm:FATE}")
+    VirtualMachine targetVM;
+
+    @Bean(name = "aeternity-configuration")
+    public AeternityServiceConfiguration aeternityConfiguration() {
+        return AeternityServiceConfiguration.configure()
+                .baseUrl(baseUrl)
+                .compilerBaseUrl(compilerBaseUrl)
+                .network(network)
+                .targetVM(targetVM)
+                .compile();
+    }
+
+    @Bean(name = "aeternity-service")
+    public AeternityService aeternityService() {
+        return new AeternityServiceFactory()
+                .getService(aeternityConfiguration());
+    }
+}
