--- provider-aeternity/src/main/java/network/arkane/provider/wallet/extraction/request/AeternityPrivateKeyExtractionRequest.java
@@ -1,18 +1,0 @@
-package network.arkane.provider.wallet.extraction.request;
-
-import lombok.Getter;
-import network.arkane.provider.chain.SecretType;
-
-@Getter
-public class AeternityPrivateKeyExtractionRequest extends ExtractionRequest {
-    private String privateKey;
-
-    public AeternityPrivateKeyExtractionRequest() {
-        super(SecretType.AETERNITY);
-    }
-
-    public AeternityPrivateKeyExtractionRequest(final String privateKey) {
-        super(SecretType.AETERNITY);
-        this.privateKey = privateKey;
-    }
-}
