+++ provider-aeternity/src/test/resources/configuration.properties
@@ -1,2 +1,4 @@
-network.arkane.aeternity.api.baseUrl=http://localhost/v2
+network.arkane.aeternity.node.api.baseUrl=http://localhost/v2
+network.arkane.aeternity.compiler.api.baseUrl=http://localhost:3080
 network.arkane.aeternity.network=DEVNET
+network.arkane.aeternity.vm=FATE
